Did minor matching of code in TwitterSignInProcess for WebApp/Campaigns. Moved muiOverrides into own file so they can be different between WebApp and Campaigns.
